Ascend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.0",
    "discord-py>=2.6.3",
    "python-dotenv>=1.1.1",
    "spotipy>=2.25.1",
    "wavelink>=3.4.1",
]
